engine: enforce explicit integer and foreign ABI contracts (#8)

+++ engine/platform/win32/win_input.cpp
@@ -1590,8 +1590,8 @@ static void IN_StartupMIDI( void ) {
 	//
 	if ( midiInOpen( &s_midiInfo.hMidiIn,
 			 in_mididevice->integer,
-			 (unsigned long)MidiInProc,
-			 (unsigned long)NULL,
+			 (DWORD)MidiInProc,
+			 (DWORD)NULL,
 			 CALLBACK_FUNCTION ) != MMSYSERR_NOERROR ) {
 		Com_DPrintf( "WARNING: could not open MIDI device %d: '%s'\n",
 			in_mididevice->integer, s_midiInfo.caps[(int)in_mididevice->value].szPname );
